Improvements

+++ setup.sh
@@ -1,4 +1,4 @@
-#!/usr/bin/bash -xe
+#!/usr/bin/bash -e
 ROOT="$(cd "$(dirname "$(readlink -f ${BASH_SOURCE[0]})")" &> /dev/null && pwd)"
 source "$ROOT/functions.sh"
 
@@ -22,51 +22,67 @@ git-get https://github.com/junegunn/fzf.git \
     ~/.fzf
 
 # Init vim
+echo "Updating vim plugins"
 vim +PlugClean! +PlugUpdate +qa > /dev/null 2>&1
 
 # Download and enable syncthing
+# It will autoupdate, so install it precisely once
 if [ ! -f ~/Programs/syncthing/syncthing ]; then
     latest-release syncthing/syncthing "syncthing-linux-amd64.*tar\.gz$" \
-        ~/Programs/syncthing.tar.gz
+            ~/Programs/syncthing.tar.gz
     tar -zxvf ~/Programs/syncthing.tar.gz
     rm ~/Programs/syncthing.tar.gz
 fi
 systemctl enable --now --user syncthing
 
 # Gyro support for yuzu and cemu
-latest-release kmicki/SteamDeckGyroDSU "SteamDeckGyroDSUSetup\.zip$" \
+try update-to-latest-release \
+        kmicki/SteamDeckGyroDSU \
+        "$(cat ~/.cache/.SteamDeckGyroDSU.version || echo)" \
+        "SteamDeckGyroDSUSetup\.zip$" \
         ~/Programs/sdgyrodsu.zip
-unzip -o ~/Programs/sdgyrodsu.zip -d ~/Programs >/dev/null;
-rm -f ~/Programs/sdgyrodsu.zip
-sed -i 's/ExecStart=.*/ExecStart=%h\/.bin\/sdgyrodsu/g' ~/Programs/SteamDeckGyroDSUSetup/sdgyrodsu.service
-sed -i 's/HOME\/sdgyrodsu/HOME\/.bin/g' ~/Programs/SteamDeckGyroDSUSetup/install.sh
 
-if groups | grep -q usbaccess; then
-    mkdir -p ~/.config/systemd/user/
-    mv ~/Programs/SteamDeckGyroDSUSetup/sdgyrodsu.service ~/.config/systemd/user/
-    mv ~/Programs/SteamDeckGyroDSUSetup/sdgyrodsu ~/.bin/
-else
-    (cd ~/Programs/SteamDeckGyroDSUSetup/; ./install.sh)
+if [ $EXIT_CODE -eq 0 ]; then
+    echo "$LATEST_VERSION" >| ~/.cache/.SteamDeckGyroDSU.version
+    unzip -o ~/Programs/sdgyrodsu.zip -d ~/Programs >/dev/null;
+    rm -f ~/Programs/sdgyrodsu.zip
+    sed -i 's/ExecStart=.*/ExecStart=%h\/.bin\/sdgyrodsu/g' ~/Programs/SteamDeckGyroDSUSetup/sdgyrodsu.service
+    sed -i 's/HOME\/sdgyrodsu/HOME\/.bin/g' ~/Programs/SteamDeckGyroDSUSetup/install.sh
+
+    if groups | grep -q usbaccess; then
+        mkdir -p ~/.config/systemd/user/
+        mv ~/Programs/SteamDeckGyroDSUSetup/sdgyrodsu.service ~/.config/systemd/user/
+        mv ~/Programs/SteamDeckGyroDSUSetup/sdgyrodsu ~/.bin/
+    else
+        (cd ~/Programs/SteamDeckGyroDSUSetup/; ./install.sh)
+    fi
+
+    rm -rf ~/Programs/SteamDeckGyroDSUSetup
+    systemctl enable --now --user sdgyrodsu
 fi
 
-rm -rf ~/Programs/SteamDeckGyroDSUSetup
-systemctl enable --now --user sdgyrodsu
-
 # Download starship
-if [ ! -f ~/.bin/starship ]; then
-    latest-release starship/starship "starship-x86_64-unknown-linux-gnu\.tar\.gz$" \
+try update-to-latest-release \
+        starship/starship \
+        "$(starship -V | awk '{print "v"$2}' || echo)" \
+        "starship-x86_64-unknown-linux-gnu\.tar\.gz$" \
         /tmp/starship.tar.gz
+
+if [ $EXIT_CODE -eq 0 ]; then 
     tar -zxvf /tmp/starship.tar.gz -C ~/.bin/
     rm /tmp/starship.tar.gz
 fi
 
 # Sunshine for streaming, not in flathub yet
-latest-release LizardByte/Sunshine "sunshine_x86_64.flatpak$" \
+try update-to-latest-release LizardByte/Sunshine \
+        "$(flatpak run dev.lizardbyte.sunshine --version | grep version | awk '{print $3}')" \
+        "sunshine_x86_64.flatpak$" \
         /tmp/sunshine.flatpak
-flatpak install --user --noninteractive --or-update \
-    /tmp/sunshine.flatpak || true
-rm /tmp/sunshine.flatpak
-
+if [ $EXIT_CODE -eq 0 ]; then 
+    flatpak install --user --noninteractive --or-update \
+        /tmp/sunshine.flatpak || true
+    rm /tmp/sunshine.flatpak
+fi
 # Install flatpaks
 flatpak install --noninteractive --or-update flathub \
     `# Apps` \
